Move MediaUploder out of index
